refactor: separate home widgets
